Retrieve and show daemon connections

+++ NervaWalletMiner/Objects/DaemonStats.cs
@@ -6,8 +6,8 @@
         public ulong YourHeight { get; set; }
         public string? NetHash { get; set; }
         public string? RunTime { get; set; }
-        public int ConnectionsIn { get; set; }
-        public int ConnectionsOut { get; set; }
+        public ulong ConnectionsIn { get; set; }
+        public ulong ConnectionsOut { get; set; }
         public string? StatusSync { get; set; }
         public string? Version { get; set; }
     }
